fp  ==>  Pike_fp Use fdlib.
Rev: src/modules/HTTPLoop/accept_and_parse.c:1.13
Rev: src/modules/HTTPLoop/accept_and_parse.h:1.6
Rev: src/modules/HTTPLoop/log.c:1.6
Rev: src/modules/HTTPLoop/requestobject.c:1.9
Rev: src/modules/HTTPLoop/util.c:1.3
Rev: src/modules/HTTPLoop/util.h:1.2
Rev: src/modules/files/file.c:1.178
Rev: src/modules/files/sendfile.c:1.39
Rev: src/modules/files/socket.c:1.49
Rev: src/modules/files/udp.c:1.11

+++ src/modules/HTTPLoop/util.c
@@ -31,7 +31,7 @@ int aap_get_time(void)
   return last_time;
 }
 
-int aap_swrite(int to, char *buf, int towrite)
+int aap_swrite(int to, char *buf, size_t towrite)
 {
   int res=0;
   int sent = 0;
